Rename DecisionModal to DeletionDecisionModal to be more accurate
