+++ frontend/src/components/DeletionDecisionModal.vue
